Table creation added.
If you call that create table method in the DB class it will attempt to
add the users table to store user data.  If the table already exists
(which it should) it will record the error to the log that the table
already exits.  No worries here though the prog will keep running.