Update user added to DB
To update a user call the updateUserData() method and pass in a user
data object with the updated information.  Make sure that you provide
all the user information fields even if it is not being updated as the
uopdate method overwrites all fields in the DB.  For example, if I were
to update a particular user's ccnum, I would get the user's data from
the db which would return a userDataObj.  I would save the obj, then
update just the ccnum field in that obj, and pass in the updated
userDataObj and the current email to update the user data.